+++ crates/swc_css_visit/src/lib.rs
@@ -716,6 +716,7 @@ define!({
     pub enum MediaInParens {
         MediaCondition(MediaCondition),
         Feature(Box<MediaFeature>),
+        GeneralEnclosed(GeneralEnclosed),
     }
 
     pub enum MediaFeature {